pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gapms"
version = "0.1.3"
description = "GAP-MS is a pipeline designed to evaluate and refine predicted gene models by integrating proteogenomic evidence from mass spectrometry with in-silico gene predictions."
authors = [{ name = "Qussai Abbas", email = "qussai.abbas@tum.de" }]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "pandas==2.2.3",
    "matplotlib==3.8.3",
    "numpy==1.24.4",
    "seaborn==0.12.2",
    "biopython==1.83",
    "scipy==1.12.0",
    "scikit-learn==1.4.1.post1",
    "xgboost==2.0.3",
    "shap==0.46.0",
    "psauron==1.0.4"
]

[project.scripts]
gapms = "gapms.main:main"
